+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 UNRELEASED
 ----------
 
+
+2022.5.4
+--------
+
 * [ [#1454](https://github.com/digitalfabrik/integreat-cms/issues/1454) ] Enable recurring events for non-expert users
 * [ [#1416](https://github.com/digitalfabrik/integreat-cms/issues/1416) ] Hide staff users from region user list
 * [ [#1483](https://github.com/digitalfabrik/integreat-cms/issues/1483) ] Add filters to admin user list